Update repositories.txt
This was an internal tool used to convert tables in UTF8-mb4 to UTF8 on newly bootstrapped development installation. Useful if you have a remote stage which does not support UTF8-mb4.

Unfortunately we did not found a comfortable way to install directly from a private repository, so we have refactored it a bit and we are going to make it public and thus easily installable.