Update repositories.txt
Recently when testing some user registration stuff it occurred to me that I could shorten my keystrokes on the db backup.  
Step 1 is this package.
Step 2 is for this package to install a plugin the firs time dbsnap is run that adds a link in the admin bar to restore immediately.